// src/app/generated/alloy.api/model/eventPermissionClaim.ts
/**
 * Alloy API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { EventPermission } from './eventPermission';


export interface EventPermissionClaim { 
    eventId?: string;
    permissions?: Array<EventPermission> | null;
}